Fixes bug with page read saving the previous one.
This was due to the fact that our wrapper is using the pageindex for most things, but when we pass a page into our activity it needs the actual page. Since the pageindex is zero-based, when we receive a message from the pdf renderer, add a page to that value.